+++ wod/admin.py
@@ -1,4 +1,5 @@
 import imp
+
 from django.contrib import admin
 
 from wod.models.characters.human import (
